Provide activities in sub modules
